Remove size_t from audio IPC code.
The change got to this size because I had to modify the surrounding code (I didn't want to just cast at the last minute).
Review URL: http://codereview.chromium.org/577006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@38192 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/audio/fake_audio_output_stream.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2009 The Chromium Authors. All rights reserved.
+// Copyright (c) 2010 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
@@ -19,7 +19,7 @@ class FakeAudioOutputStream : public AudioOutputStream {
   static AudioOutputStream* MakeFakeStream();
   static FakeAudioOutputStream* GetLastFakeStream();
 
-  virtual bool Open(size_t packet_size);
+  virtual bool Open(uint32 packet_size);
   virtual void Start(AudioSourceCallback* callback);
   virtual void Stop();
   virtual void SetVolume(double volume);
@@ -40,7 +40,7 @@ class FakeAudioOutputStream : public AudioOutputStream {
   double volume_;
   AudioSourceCallback* callback_;
   scoped_array<char> buffer_;
-  size_t packet_size_;
+  uint32 packet_size_;
 
   DISALLOW_COPY_AND_ASSIGN(FakeAudioOutputStream);
 };
